add store macro to generate extension traits
