+++ examples/todomvc_store.rs
@@ -21,6 +21,42 @@ struct TodoState {
     filter: FilterState,
 }
 
+// We can also add custom methods to the store by using the `store` attribute on an impl block.
+// The store attribute turns the impl block into an extension trait for Store<TodoState>.
+// Methods that take &self will automatically get a bound that Lens: Readable<Target = TodoState>
+// Methods that take &mut self will automatically get a bound that Lens: Writable<Target = TodoState>
+#[store]
+impl<Lens> Store<TodoState, Lens> {
+    fn active_items(&self) -> Vec<u32> {
+        let filter = self.filter().cloned();
+        let mut active_ids: Vec<u32> = self
+            .todos()
+            .iter()
+            .filter_map(|(id, item)| item.active(filter).then_some(*id))
+            .collect();
+        active_ids.sort_unstable();
+        active_ids
+    }
+
+    fn incomplete_count(&self) -> usize {
+        self.todos()
+            .values()
+            .filter(|item| item.incomplete())
+            .count()
+    }
+
+    fn toggle_all(&mut self) {
+        let check = self.incomplete_count() != 0;
+        for item in self.todos().values() {
+            item.checked().set(check);
+        }
+    }
+
+    fn has_todos(&self) -> bool {
+        !self.todos().is_empty()
+    }
+}
+
 #[derive(PartialEq, Eq, Clone, Copy, Debug)]
 enum FilterState {
     All,
@@ -43,58 +79,49 @@ impl TodoItem {
     }
 }
 
+#[store]
+impl<Lens> Store<TodoItem, Lens> {
+    fn complete(&self) -> bool {
+        self.checked().cloned()
+    }
+
+    fn incomplete(&self) -> bool {
+        !self.complete()
+    }
+
+    fn active(&self, filter: FilterState) -> bool {
+        match filter {
+            FilterState::All => true,
+            FilterState::Active => self.incomplete(),
+            FilterState::Completed => self.complete(),
+        }
+    }
+}
+
 fn main() {
     dioxus::launch(app);
 }
 
 fn app() -> Element {
     // We store the state of our todo list in a store to use throughout the app.
-    let todos = use_store(|| TodoState {
+    let mut todos = use_store(|| TodoState {
         todos: HashMap::new(),
         filter: FilterState::All,
     });
 
-    // We can use the methods generated by the store macro to zoom into the fields of the struct.
-    // This gives us a Store<HashMap<u32, TodoItem>> for the todos and
-    // a Store<FilterState> for the filter.
-    let todo_list = todos.todos();
-    let filter = todos.filter();
-
     // We use a simple memo to calculate the number of active todos.
     // Whenever the todos change, the active_todo_count will be recalculated.
-    let active_todo_count = use_memo(move || {
-        todo_list
-            .values()
-            .filter(|item| !item.checked().cloned())
-            .count()
-    });
+    let active_todo_count = use_memo(move || todos.incomplete_count());
 
     // We use a memo to filter the todos based on the current filter state.
     // Whenever the todos or filter change, the filtered_todos will be recalculated.
     // Note that we're only storing the IDs of the todos, not the todos themselves.
-    let filtered_todos = use_memo(move || {
-        let mut filtered_todos = todo_list
-            .iter()
-            .filter(|(_, item)| match filter() {
-                FilterState::All => true,
-                FilterState::Active => !item.checked().cloned(),
-                FilterState::Completed => item.checked().cloned(),
-            })
-            .map(|(i, _)| i)
-            .collect::<Vec<_>>();
-
-        filtered_todos.sort_unstable();
-
-        filtered_todos
-    });
+    let filtered_todos = use_memo(move || todos.active_items());
 
     // Toggle all the todos to the opposite of the current state.
     // If all todos are checked, uncheck them all. If any are unchecked, check them all.
     let toggle_all = move |_| {
-        let check = active_todo_count() != 0;
-        for item in todo_list.values() {
-            item.checked().set(check);
-        }
+        todos.toggle_all();
     };
 
     rsx! {
@@ -102,7 +129,7 @@ fn app() -> Element {
         section { class: "todoapp",
             TodoHeader { todos }
             section { class: "main",
-                if !todo_list.is_empty() {
+                if todos.has_todos() {
                     input {
                         id: "toggle-all",
                         class: "toggle-all",
@@ -122,7 +149,7 @@ fn app() -> Element {
                 }
 
                 // We only show the footer if there are todos.
-                if !todo_list.is_empty() {
+                if todos.has_todos() {
                     ListFooter { active_todo_count, todos }
                 }
             }
@@ -241,8 +268,7 @@ fn ListFooter(mut todos: Store<TodoState>, active_todo_count: ReadSignal<usize>)
     // We use a memo to calculate whether we should show the "Clear completed" button.
     // This will recompute whenever the number of todos change or the checked state of an existing
     // todo changes
-    let show_clear_completed =
-        use_memo(move || todos.todos().values().any(|todo| todo.checked().cloned()));
+    let show_clear_completed = use_memo(move || todos.todos().values().any(|todo| todo.complete()));
     let mut filter = todos.filter();
 
     rsx! {
