+++ packages/stores-macro/Cargo.toml
@@ -12,7 +12,7 @@ description = "Server function macros for Dioxus"
 [dependencies]
 proc-macro2 = { workspace = true }
 quote = { workspace = true }
-syn = { workspace = true }
+syn = { workspace = true, features = ["full"] }
 convert_case = { workspace = true }
 
 [dev-dependencies]
